Qode/requirements.txt
sqlite3
flask 
flask_session
werkzeug.security
groq
os